Fix flow annotations
